legal_case_assistant_agent/legal_case_assistant_agent/requirements.txt
google-cloud-logging>=3.15.0
python-dotenv>=1.2.2
toolbox-adk>=1.0.0
toolbox-core>=1.0.0
